refactor: extract execution ports
