+++ crates/cli/src/application.rs
@@ -65,6 +65,13 @@ use crate::{
     },
 };
 
+mod execution;
+
+use execution::{
+    InstructionEntry, InstructionPort, InstructionWrite, ManagedSkillAction, ManagedSkillEntry,
+    ManagedSkillPort, StateExecutionJournal,
+};
+
 pub(crate) struct StatusApplication<'a> {
     pub(crate) config: &'a dyn ConfigRepository,
     pub(crate) inventory: &'a dyn InventoryRepository,
@@ -102,534 +109,6 @@ pub(crate) struct SkillInstallRequest<'a> {
 /// result through the validated plan, seeds only explicitly planned resources,
 /// updates exact resource records, and publishes atomically through the
 /// repository port.
-pub(crate) struct StateExecutionJournal<'a> {
-    pub(crate) plan: &'a Plan,
-    pub(crate) state: &'a dyn StateRepository,
-    pub(crate) seeds: BTreeMap<ResourceKey, ResourceState>,
-}
-
-impl ExecutionJournal for StateExecutionJournal<'_> {
-    fn record(&self, result: &OperationResult) -> Result<(), ExecutionError> {
-        let operation = self.plan.get(result.operation_id()).ok_or_else(|| {
-            ExecutionError::journal_failure(
-                skilltap_core::domain::EvidenceCode::new("state.operation_unknown")
-                    .expect("static evidence code is valid"),
-                skilltap_core::domain::EvidenceDetail::new(
-                    "The state journal received an operation outside the validated plan.",
-                )
-                .expect("static evidence detail is valid"),
-            )
-        })?;
-        let resource = operation.selector().resource();
-        let current = self.state.load().map_err(|_| {
-            ExecutionError::journal_failure(
-                skilltap_core::domain::EvidenceCode::new("state.load_failed")
-                    .expect("static evidence code is valid"),
-                skilltap_core::domain::EvidenceDetail::new(
-                    "The state document could not be loaded for journaling.",
-                )
-                .expect("static evidence detail is valid"),
-            )
-        })?;
-        let current = match current {
-            DocumentState::Present(current) => current,
-            DocumentState::Missing => skilltap_core::storage::StateDocument::new(
-                skilltap_core::storage::STATE_SCHEMA_VERSION,
-                [],
-                [],
-                None,
-                None,
-                None,
-            )
-            .map_err(|_| {
-                ExecutionError::journal_failure(
-                    skilltap_core::domain::EvidenceCode::new("state.seed_invalid")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The seed state for the operation was invalid.",
-                    )
-                    .expect("static evidence detail is valid"),
-                )
-            })?,
-        };
-        let current = if current.resources().contains_key(resource) {
-            if let Some(seed) = self.seeds.get(resource) {
-                current.refresh_resource_state(seed.clone()).map_err(|_| {
-                    ExecutionError::journal_failure(
-                        skilltap_core::domain::EvidenceCode::new("state.seed_refresh_failed")
-                            .expect("static evidence code is valid"),
-                        skilltap_core::domain::EvidenceDetail::new(
-                            "The existing resource metadata could not be refreshed safely.",
-                        )
-                        .expect("static evidence detail is valid"),
-                    )
-                })?
-            } else {
-                current
-            }
-        } else if let Some(seed) = self.seeds.get(resource) {
-            current.with_resource_state(seed.clone()).map_err(|_| {
-                ExecutionError::journal_failure(
-                    skilltap_core::domain::EvidenceCode::new("state.seed_conflict")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The operation resource could not be seeded in state.",
-                    )
-                    .expect("static evidence detail is valid"),
-                )
-            })?
-        } else {
-            return Err(ExecutionError::journal_failure(
-                skilltap_core::domain::EvidenceCode::new("state.resource_missing")
-                    .expect("static evidence code is valid"),
-                skilltap_core::domain::EvidenceDetail::new(
-                    "The operation resource is not present in state.",
-                )
-                .expect("static evidence detail is valid"),
-            ));
-        };
-        let at = Timestamp::from_system_time(std::time::SystemTime::now()).map_err(|_| {
-            ExecutionError::journal_failure(
-                skilltap_core::domain::EvidenceCode::new("state.clock_invalid")
-                    .expect("static evidence code is valid"),
-                skilltap_core::domain::EvidenceDetail::new(
-                    "The operation timestamp could not be recorded.",
-                )
-                .expect("static evidence detail is valid"),
-            )
-        })?;
-        let next = current
-            .with_operation_result(resource, at, result.clone())
-            .map_err(|_| {
-                ExecutionError::journal_failure(
-                    skilltap_core::domain::EvidenceCode::new("state.resource_unavailable")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The operation resource could not be journaled in state.",
-                    )
-                    .expect("static evidence detail is valid"),
-                )
-            })?;
-        self.state.replace(&next).map_err(|_| {
-            ExecutionError::journal_failure(
-                skilltap_core::domain::EvidenceCode::new("state.publish_failed")
-                    .expect("static evidence code is valid"),
-                skilltap_core::domain::EvidenceDetail::new(
-                    "The state journal could not be published atomically.",
-                )
-                .expect("static evidence detail is valid"),
-            )
-        })
-    }
-}
-
-struct ManagedSkillPort<'a> {
-    filesystem: &'a dyn DirectoryTreeFileSystem,
-    entries: BTreeMap<OperationId, ManagedSkillEntry>,
-}
-
-struct ManagedSkillEntry {
-    root: AbsolutePath,
-    destination: skilltap_core::domain::RelativeArtifactPath,
-    tree: ArtifactTree,
-    backup_tree: Option<ArtifactTree>,
-    action: ManagedSkillAction,
-    expected_identity: Option<skilltap_core::runtime::DirectoryIdentity>,
-    owner: Option<ResourceKey>,
-    config_root: Option<AbsolutePath>,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ManagedSkillAction {
-    Install,
-    Replace,
-    Remove,
-}
-
-impl ExecutionPort for ManagedSkillPort<'_> {
-    fn revalidate(&self, plan: &Plan) -> Result<(), ExecutionError> {
-        for (_, operation) in plan.iter() {
-            if !matches!(
-                operation.action(),
-                OperationAction::SkillInstall | OperationAction::SkillRemove
-            ) {
-                continue;
-            }
-            let Some(entry) = self.entries.get(operation.id()) else {
-                return Err(ExecutionError::revalidation(
-                    skilltap_core::domain::EvidenceCode::new("managed.skill_request_missing")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The managed skill adapter did not receive a request for a planned operation.",
-                    )
-                    .expect("static evidence detail is valid"),
-                ));
-            };
-            let expected = AbsolutePath::new(format!(
-                "{}/{}",
-                entry.root.as_str(),
-                entry.destination.as_str()
-            ))
-            .map_err(|_| {
-                ExecutionError::revalidation(
-                    skilltap_core::domain::EvidenceCode::new("managed.skill_path_invalid")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The managed skill destination could not be represented safely.",
-                    )
-                    .expect("static evidence detail is valid"),
-                )
-            })?;
-            if !operation
-                .affected_surfaces()
-                .iter()
-                .any(|surface| surface.path() == Some(&expected))
-            {
-                return Err(ExecutionError::revalidation(
-                    skilltap_core::domain::EvidenceCode::new("managed.skill_surface_mismatch")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The managed skill destination no longer matches the validated operation.",
-                    )
-                    .expect("static evidence detail is valid"),
-                ));
-            }
-        }
-        Ok(())
-    }
-
-    fn apply(
-        &self,
-        operation: &skilltap_core::domain::Operation,
-    ) -> Result<OperationOutcome, ExecutionError> {
-        let Some(entry) = self.entries.get(operation.id()) else {
-            return Err(ExecutionError::revalidation(
-                skilltap_core::domain::EvidenceCode::new("managed.skill_request_missing")
-                    .expect("static evidence code is valid"),
-                skilltap_core::domain::EvidenceDetail::new(
-                    "The managed skill adapter did not receive a request for a planned operation.",
-                )
-                .expect("static evidence detail is valid"),
-            ));
-        };
-        if entry.action == ManagedSkillAction::Remove {
-            let Some(expected) = entry.expected_identity else {
-                return Err(managed_skill_apply_failure(
-                    "The managed skill removal did not include an owned directory identity.",
-                ));
-            };
-            self.filesystem
-                .remove_tree_no_follow(&entry.root, &entry.destination, expected)
-                .map(|_| OperationOutcome::Applied)
-                .map_err(|_| {
-                    managed_skill_apply_failure(
-                        "The managed skill tree could not be removed safely.",
-                    )
-                })
-        } else if entry.action == ManagedSkillAction::Replace {
-            let Some(expected) = entry.expected_identity else {
-                return Err(managed_skill_apply_failure(
-                    "The managed skill replacement did not include an owned directory identity.",
-                ));
-            };
-            let Some(owner) = &entry.owner else {
-                return Err(managed_skill_apply_failure(
-                    "The managed skill replacement did not include an ownership record.",
-                ));
-            };
-            let Some(config_root) = &entry.config_root else {
-                return Err(managed_skill_apply_failure(
-                    "The managed skill replacement did not include a backup root.",
-                ));
-            };
-            let repository = skilltap_core::storage::FileManagedArtifactRepository::new(
-                self.filesystem,
-                config_root.clone(),
-            )
-            .map_err(|_| {
-                managed_skill_apply_failure(
-                    "The managed skill backup repository could not be opened.",
-                )
-            })?;
-            let Some(backup_tree) = &entry.backup_tree else {
-                return Err(managed_skill_apply_failure(
-                    "The managed skill replacement did not include the previous tree.",
-                ));
-            };
-            let backup = repository.backup(owner, backup_tree).map_err(|_| {
-                managed_skill_apply_failure(
-                    "The existing skill tree could not be backed up safely.",
-                )
-            })?;
-            self.filesystem
-                .remove_tree_no_follow(&entry.root, &entry.destination, expected)
-                .map_err(|_| {
-                    managed_skill_apply_failure(
-                        "The existing skill tree could not be removed safely.",
-                    )
-                })?;
-            match self.filesystem.publish_tree_no_follow(
-                &entry.root,
-                &entry.destination,
-                entry.tree.files(),
-            ) {
-                Ok(skilltap_core::runtime::DirectoryPublishOutcome::Published(_)) => {
-                    Ok(OperationOutcome::Applied)
-                }
-                Ok(skilltap_core::runtime::DirectoryPublishOutcome::AlreadyExists) => {
-                    Ok(OperationOutcome::NoChange)
-                }
-                Err(_) => {
-                    let _ = self.filesystem.publish_tree_no_follow(
-                        &entry.root,
-                        &entry.destination,
-                        backup_tree.files(),
-                    );
-                    let _ = backup;
-                    Err(managed_skill_apply_failure(
-                        "The replacement skill tree could not be published after backup.",
-                    ))
-                }
-            }
-        } else {
-            match self
-                .filesystem
-                .publish_tree_no_follow(&entry.root, &entry.destination, entry.tree.files())
-                .map_err(|_| {
-                    managed_skill_apply_failure("The managed skill tree could not be published.")
-                })? {
-                skilltap_core::runtime::DirectoryPublishOutcome::Published(_) => {
-                    Ok(OperationOutcome::Applied)
-                }
-                skilltap_core::runtime::DirectoryPublishOutcome::AlreadyExists => {
-                    let (_, files) = self
-                        .filesystem
-                        .load_tree_no_follow(&entry.root, &entry.destination)
-                        .map_err(|_| {
-                            managed_skill_apply_failure(
-                                "The existing managed skill tree could not be re-read safely.",
-                            )
-                        })?;
-                    let current = ArtifactTree::new(
-                        files
-                            .into_iter()
-                            .map(|(path, bytes)| (path.as_str().to_owned(), bytes)),
-                    )
-                    .map_err(|_| {
-                        managed_skill_apply_failure(
-                            "The existing managed skill tree had an invalid shape.",
-                        )
-                    })?;
-                    if current == entry.tree {
-                        Ok(OperationOutcome::NoChange)
-                    } else {
-                        Err(managed_skill_apply_failure(
-                            "The managed skill destination changed before publication.",
-                        ))
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn managed_skill_apply_failure(detail: &'static str) -> ExecutionError {
-    ExecutionError::apply_failure(skilltap_core::domain::AttentionReason::operation_failed(
-        skilltap_core::domain::EvidenceCode::new("managed.skill_publish_failed")
-            .expect("static evidence code is valid"),
-        skilltap_core::domain::EvidenceDetail::new(detail)
-            .expect("static evidence detail is valid"),
-    ))
-}
-
-enum InstructionWrite {
-    Canonical,
-    Symlink { target: RelativeSymlinkTarget },
-    Import { contents: Vec<u8> },
-    Remove,
-}
-
-struct InstructionPort<'a> {
-    filesystem: &'a dyn FileSystem,
-    entries: BTreeMap<OperationId, InstructionEntry>,
-}
-
-struct InstructionEntry {
-    path: AbsolutePath,
-    write: InstructionWrite,
-    action: OperationAction,
-    backup: Option<AbsolutePath>,
-}
-
-impl ExecutionPort for InstructionPort<'_> {
-    fn revalidate(&self, plan: &Plan) -> Result<(), ExecutionError> {
-        for (_, operation) in plan.iter() {
-            if !matches!(
-                operation.action(),
-                OperationAction::InstructionSetup | OperationAction::InstructionRepair
-            ) {
-                continue;
-            }
-            let Some(entry) = self.entries.get(operation.id()) else {
-                return Err(ExecutionError::revalidation(
-                    skilltap_core::domain::EvidenceCode::new("instructions.request_missing")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The instruction adapter did not receive a request for a planned operation.",
-                    )
-                    .expect("static evidence detail is valid"),
-                ));
-            };
-            if entry.action != operation.action() {
-                return Err(ExecutionError::revalidation(
-                    skilltap_core::domain::EvidenceCode::new("instructions.action_mismatch")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The instruction operation action no longer matches the validated adapter entry.",
-                    )
-                    .expect("static evidence detail is valid"),
-                ));
-            }
-            if !operation
-                .affected_surfaces()
-                .iter()
-                .any(|surface| surface.path() == Some(&entry.path))
-            {
-                return Err(ExecutionError::revalidation(
-                    skilltap_core::domain::EvidenceCode::new("instructions.surface_mismatch")
-                        .expect("static evidence code is valid"),
-                    skilltap_core::domain::EvidenceDetail::new(
-                        "The instruction destination no longer matches the validated operation.",
-                    )
-                    .expect("static evidence detail is valid"),
-                ));
-            }
-        }
-        Ok(())
-    }
-
-    fn apply(
-        &self,
-        operation: &skilltap_core::domain::Operation,
-    ) -> Result<OperationOutcome, ExecutionError> {
-        let Some(entry) = self.entries.get(operation.id()) else {
-            return Err(ExecutionError::revalidation(
-                skilltap_core::domain::EvidenceCode::new("instructions.request_missing")
-                    .expect("static evidence code is valid"),
-                skilltap_core::domain::EvidenceDetail::new(
-                    "The instruction adapter did not receive a request for a planned operation.",
-                )
-                .expect("static evidence detail is valid"),
-            ));
-        };
-        if matches!(&entry.write, InstructionWrite::Remove) {
-            if let Some(backup) = &entry.backup {
-                let backup_parent = backup
-                    .as_str()
-                    .rsplit_once('/')
-                    .map(|(parent, _)| parent)
-                    .and_then(|parent| AbsolutePath::new(parent).ok())
-                    .ok_or_else(|| {
-                        instruction_apply_failure("The instruction backup path is invalid.")
-                    })?;
-                self.filesystem
-                    .create_directory_all(&backup_parent)
-                    .map_err(|_| {
-                        instruction_apply_failure(
-                            "The existing instruction bridge could not be backed up safely.",
-                        )
-                    })?;
-                self.filesystem
-                    .copy_recoverable(&entry.path, backup)
-                    .map_err(|_| {
-                        instruction_apply_failure(
-                            "The existing instruction bridge could not be backed up safely.",
-                        )
-                    })?;
-            }
-            self.filesystem.remove(&entry.path).map_err(|_| {
-                instruction_apply_failure(
-                    "The duplicate instruction bridge could not be removed safely.",
-                )
-            })?;
-            return Ok(OperationOutcome::Applied);
-        }
-        let parent = entry
-            .path
-            .as_str()
-            .rsplit_once('/')
-            .map(|(parent, _)| parent)
-            .and_then(|parent| AbsolutePath::new(parent).ok())
-            .ok_or_else(|| instruction_apply_failure("The instruction parent path is invalid."))?;
-        self.filesystem.create_directory_all(&parent).map_err(|_| {
-            instruction_apply_failure("The instruction parent directory could not be created.")
-        })?;
-        if let Some(backup) = &entry.backup {
-            let backup_parent = backup
-                .as_str()
-                .rsplit_once('/')
-                .map(|(parent, _)| parent)
-                .and_then(|parent| AbsolutePath::new(parent).ok())
-                .ok_or_else(|| {
-                    instruction_apply_failure("The instruction backup path is invalid.")
-                })?;
-            self.filesystem
-                .create_directory_all(&backup_parent)
-                .map_err(|_| {
-                    instruction_apply_failure(
-                        "The instruction backup directory could not be created.",
-                    )
-                })?;
-            self.filesystem
-                .copy_recoverable(&entry.path, backup)
-                .map_err(|_| {
-                    instruction_apply_failure(
-                        "The existing instruction bridge could not be backed up safely.",
-                    )
-                })?;
-            self.filesystem.remove(&entry.path).map_err(|_| {
-                instruction_apply_failure(
-                    "The existing instruction bridge could not be replaced safely.",
-                )
-            })?;
-        }
-        match &entry.write {
-            InstructionWrite::Canonical => {
-                self.filesystem
-                    .atomic_write(&entry.path, &[])
-                    .map_err(|_| {
-                        instruction_apply_failure(
-                            "The canonical instruction file could not be created.",
-                        )
-                    })?
-            }
-            InstructionWrite::Symlink { target } => self
-                .filesystem
-                .create_relative_symlink(target, &entry.path)
-                .map_err(|_| {
-                    instruction_apply_failure("The instruction bridge could not be created.")
-                })?,
-            InstructionWrite::Import { contents } => self
-                .filesystem
-                .atomic_write(&entry.path, contents)
-                .map_err(|_| {
-                    instruction_apply_failure("The instruction import bridge could not be created.")
-                })?,
-            InstructionWrite::Remove => unreachable!("remove entries return before publication"),
-        }
-        Ok(OperationOutcome::Applied)
-    }
-}
-
-fn instruction_apply_failure(detail: &'static str) -> ExecutionError {
-    ExecutionError::apply_failure(skilltap_core::domain::AttentionReason::operation_failed(
-        skilltap_core::domain::EvidenceCode::new("instructions.publish_failed")
-            .expect("static evidence code is valid"),
-        skilltap_core::domain::EvidenceDetail::new(detail)
-            .expect("static evidence detail is valid"),
-    ))
-}
-
 impl StatusApplication<'_> {
     /// Build a fresh reconciliation plan from the current documents and
     /// bounded native observation.  The desired inventory is the source of
